Refactor log retrieval CLI command and update tests
- Moved the `GetLogsHandler` and related classes to a new `logs.py` module for better organization.
- Updated the CLI command to directly use `logs` instead of `get logs`, simplifying the command structure.
- Adjusted unit tests to reflect the new command structure and updated import paths accordingly.
- Enhanced test coverage for log retrieval functionality, ensuring robust error handling and user interaction.
