+++ tests/test_get_logs.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest.mock import MagicMock, patch
 
-from weni_cli.commands.logs.get import GetLogsHandler
+from weni_cli.commands.logs import GetLogsHandler
 from click.testing import CliRunner
 from weni_cli.cli import cli
 
@@ -9,7 +9,7 @@ from weni_cli.cli import cli
 @pytest.fixture
 def mock_cli_client():
     """Mock the CLIClient instance."""
-    with patch("weni_cli.commands.logs.get.CLIClient") as mock_client_class:
+    with patch("weni_cli.commands.logs.CLIClient") as mock_client_class:
         mock_client = MagicMock()
         mock_client_class.return_value = mock_client
         yield mock_client
@@ -18,7 +18,7 @@ def mock_cli_client():
 @pytest.fixture
 def mock_formatter():
     """Mock the Formatter instance."""
-    with patch("weni_cli.commands.logs.get.Formatter") as mock_formatter_class:
+    with patch("weni_cli.commands.logs.Formatter") as mock_formatter_class:
         mock_formatter_instance = MagicMock()
         mock_formatter_class.return_value = mock_formatter_instance
         yield mock_formatter_instance
@@ -27,7 +27,7 @@ def mock_formatter():
 @pytest.fixture
 def mock_console():
     """Mock the Console instance."""
-    with patch("weni_cli.commands.logs.get.Console") as mock_console_class:
+    with patch("weni_cli.commands.logs.Console") as mock_console_class:
         mock_console = MagicMock()
         mock_console_class.return_value = mock_console
         yield mock_console
@@ -36,7 +36,7 @@ def mock_console():
 @pytest.fixture
 def mock_print():
     """Mock the rich print function."""
-    with patch("weni_cli.commands.logs.get.print") as mock_print:
+    with patch("weni_cli.commands.logs.print") as mock_print:
         yield mock_print
 
 
@@ -113,14 +113,14 @@ def test_cli_get_logs_success(mocker):
     """Test CLI invocation of the get logs command with successful response."""
     # Mock the GetLogsHandler
     mock_handler = MagicMock()
-    mocker.patch("weni_cli.commands.logs.get.GetLogsHandler", return_value=mock_handler)
+    mocker.patch("weni_cli.commands.logs.GetLogsHandler", return_value=mock_handler)
 
     # Create CLI runner
     runner = CliRunner()
 
     # Invoke the CLI command
     result = runner.invoke(
-        cli, ["get", "logs", "--agent", "test-agent", "--tool", "test-tool"]
+        cli, ["logs", "--agent", "test-agent", "--tool", "test-tool"]
     )
 
     # Verify CLI command executed successfully
@@ -136,7 +136,7 @@ def test_cli_get_logs_with_time_range_and_pattern(mocker):
     """Test CLI invocation with time range and pattern parameters."""
     # Mock the GetLogsHandler
     mock_handler = MagicMock()
-    mocker.patch("weni_cli.commands.logs.get.GetLogsHandler", return_value=mock_handler)
+    mocker.patch("weni_cli.commands.logs.GetLogsHandler", return_value=mock_handler)
 
     # Create CLI runner
     runner = CliRunner()
@@ -144,7 +144,7 @@ def test_cli_get_logs_with_time_range_and_pattern(mocker):
     # Invoke the CLI command with time range and pattern
     result = runner.invoke(
         cli, [
-            "get", "logs",
+            "logs",
             "--agent", "test-agent",
             "--tool", "test-tool",
             "--start-time", "2023-01-01T00:00:00",
@@ -171,14 +171,14 @@ def test_cli_get_logs_error(mocker):
     # Mock the GetLogsHandler to raise an exception
     mock_handler = MagicMock()
     mock_handler.get_logs.side_effect = Exception("Test error")
-    mocker.patch("weni_cli.commands.logs.get.GetLogsHandler", return_value=mock_handler)
+    mocker.patch("weni_cli.commands.logs.GetLogsHandler", return_value=mock_handler)
 
     # Create CLI runner
     runner = CliRunner()
 
     # Invoke the CLI command
     result = runner.invoke(
-        cli, ["get", "logs", "--agent", "test-agent", "--tool", "test-tool"]
+        cli, ["logs", "--agent", "test-agent", "--tool", "test-tool"]
     )
 
     # Verify CLI command executed successfully but handled the error
@@ -192,7 +192,7 @@ def test_cli_get_logs_missing_required_args():
     runner = CliRunner()
 
     # Invoke the CLI command without required agent
-    result = runner.invoke(cli, ["get", "logs", "--tool", "test-tool"])
+    result = runner.invoke(cli, ["logs", "--tool", "test-tool"])
 
     # Verify CLI command failed due to missing required argument
     assert result.exit_code != 0
@@ -200,7 +200,7 @@ def test_cli_get_logs_missing_required_args():
     assert "--agent" in result.output
 
     # Invoke the CLI command without required tool
-    result = runner.invoke(cli, ["get", "logs", "--agent", "test-agent"])
+    result = runner.invoke(cli, ["logs", "--agent", "test-agent"])
 
     # Verify CLI command failed due to missing required argument
     assert result.exit_code != 0
@@ -217,7 +217,7 @@ def test_get_logs_invalid_pattern(mock_cli_client, mock_formatter, mock_console)
     mock_cli_client.get_tool_logs.assert_not_called()
 
 
-@patch("weni_cli.commands.logs.get.Confirm.ask")
+@patch("weni_cli.commands.logs.Confirm.ask")
 def test_get_logs_pagination_no_more_logs(mock_confirm_ask, mock_cli_client, mock_formatter, mock_console, mock_print):
     """Test pagination where the second page has no logs."""
     # Mock Confirm.ask to continue fetching
@@ -259,8 +259,8 @@ def test_get_logs_pagination_no_more_logs(mock_confirm_ask, mock_cli_client, moc
     mock_confirm_ask.assert_called_once_with("Fetch more logs?", default=True)
 
 
-@patch("weni_cli.commands.logs.get.Confirm.ask")
-@patch("weni_cli.commands.logs.get.CLIClient")
+@patch("weni_cli.commands.logs.Confirm.ask")
+@patch("weni_cli.commands.logs.CLIClient")
 def test_cli_get_logs_pagination_user_declines(mock_cli_client_class, mock_confirm_ask, mock_formatter, mock_console):
     """Test CLI invocation with pagination where the user declines fetching more logs."""
     # Mock Confirm.ask to stop fetching
@@ -283,7 +283,7 @@ def test_cli_get_logs_pagination_user_declines(mock_cli_client_class, mock_confi
     # Invoke the CLI command
     # We don't need mocker here as we patch dependencies directly
     result = runner.invoke(
-        cli, ["get", "logs", "--agent", "test-agent-cli", "--tool", "test-tool-cli"]
+        cli, ["logs", "--agent", "test-agent-cli", "--tool", "test-tool-cli"]
     )
 
     # Verify CLI command executed successfully
